requirements.txt
fastapi
uvicorn
yfinance
alpha_vantage
requests
pydantic
python-dotenv
openai
alpaca-trade-api
google-generativeai
